update DetailsOfInstallationService.kt.kt, update ResponsibleAdultService.kt, update VariationService.kt. Move UpdateDetailsOfInstallationDto.kt
