+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppselectronicmonitoringcreateanorderapi/service/VariationService.kt
@@ -34,6 +34,6 @@ class VariationService : OrderSectionServiceBase() {
       )
     }
 
-    return orderRepo.save(order).variationDetails!!
+    return updateLastUpdatedByAndSaveOrder(order).variationDetails!!
   }
 }
